Extract Case and When from core.py
