+++ magia/factory.py
@@ -10,6 +10,7 @@ import typing
 
 if typing.TYPE_CHECKING:
     from .comb_ops import Operation
+    from .comb_select import Case, When
     from .constant import Constant
     from .core import Signal
     from .data_struct import OPType
@@ -33,6 +34,18 @@ def sv_constant(value: None | int | bytes, width: int, signed: bool = False) -> 
 
 def create_comb_op(op_type: OPType, x: Signal, y: None | Signal | slice | int | bytes) -> Operation:
     return Operation.create(op_type, x, y)
+
+
+def create_when(condition: Signal, if_true: Signal, if_false: None | Signal | int | bytes, **kwargs) -> When:
+    return When(condition, if_true, if_false, **kwargs)
+
+
+def create_case(
+        selector: Signal, cases: dict[int, Signal | int],
+        default: None | Signal | int = None,
+        **kwargs
+) -> Case:
+    return Case(selector, cases, default, **kwargs)
 
 
 def register(
@@ -62,10 +75,14 @@ def register(
 
 def deferred_imports():
     from .comb_ops import Operation as OperationImported
+    from .comb_select import Case as CaseImported
+    from .comb_select import When as WhenImported
     from .constant import Constant as ConstantImported
     from .register import Register as RegisterImported
     globals().update({
         "Constant": ConstantImported,
         "Operation": OperationImported,
         "Register": RegisterImported,
+        "Case": CaseImported,
+        "When": WhenImported
     })
